Fix security vulnerability in ffi
I upgraded sass-rails v5.0.6 -> v5.0.7 to trigger upgrade to sass and
thence ffi v1.9.18 -> v1.9.25.

CVE-2018-1000201 [1]: ruby-ffi version 1.9.23 and earlier has a DLL
loading issue which can be hijacked on Windows OS, when a Symbol is used
as DLL name instead of a String This vulnerability appears to have been
fixed in v1.9.24 and later.

[1]: https://nvd.nist.gov/vuln/detail/CVE-2018-1000201