Redirect reqs for recap docs to freerange GH Pages site
* This is far from ideal, but the best option I can come up with. My
motivation is to remove the need to serve the recap documentation from
the Apache instance running on our Linode VPS.

* This is achieved by removing the relevant Apache config and adding
redirects in the Rails routes config to send requests to the GitHub
Pages version of the docs.

Reasons why this is not ideal:

* freerange/recap is not the canonical repo, nor therefore the
canonical docs, so we might end up having to redirect again.

* The master branch is many releases behind the latest gem release which
is in the 1.x branch. However, I believe we've been publishing docs from
master on gofreerange.com/recap/docs for some time!

* It's not ideal that Rails is doing the redirecting, but there is
precedent for this in the redirects for the exciting.io-related URLs.
Also it means it should be easier to move the site to Heroku.