Move public/assets -> public/media
The public/assets directory is special in a Rails app in that its used
to store precompiled assets from the asset pipeline. I'm about to make a
change related to the Rails upgrade which will instruct git to ignore
files in this directory, so we need to move them somewhere safer!