Add a first stab at content for 2018 internal/client projects
Please feel free to suggest changes or additions to this content,
particularly if I've misrepresented something or said something
commercially sensitive.

These will appear on the projects page [1] which is linked to as "Fuller
list of projects" under the "Significant projects" section [2] on the
home page.

We might want to consider updating the this "Significant projects"
section to include one or two of our more recent projects.

[1]: https://gofreerange.com/projects
[2]: https://gofreerange.com/#work