Make atom feed available via /GoFreeRangeBlog path
We want to serve the atom feed directly rather than via FeedBurner,
because the latter does not handle SSL requests. I plan to add the
feeds.gofreerange.com sub-domain to our Heroku app along with an SSL
certificate. This should mean that we can serve
https://feeds.gofreerange.com/GoFreeRangeBlog directly from the Rails
app.

Here are some more detailed notes take from the relevant Trello card:

Previously (before the move to SSL) the link on the website worked liked
this:

http://feeds.gofreerange.com/GoFreeRangeBlog ->
http://1mcuj68.feedproxy.ghs.google.com/GoFreeRangeBlog ->
http://feeds.feedburner.com/GoFreeRangeBlog ->
http://gofreerange.com/feed.xml

Note that none of these are redirections, rather they are either DNS
resolution or proxying (in the latter case).

Currently (after the move to SSL) the link on the website tries to do
the following but fails:

https://feeds.gofreerange.com/GoFreeRangeBlog ->
https://1mcuj68.feedproxy.ghs.google.com/GoFreeRangeBlog ->
https://feeds.feedburner.com/GoFreeRangeBlog ->
http://gofreerange.com/feed.xml

I think it fails because feeds.feedburner.com/GoFreeRangeBlog is not
available over SSL:

    $ curl -I https://feeds.gofreerange.com/GoFreeRangeBlog
    curl: (35) LibreSSL SSL_connect: SSL_ERROR_SYSCALL in connection to feeds.gofreerange.com:443

Changing the link on the website to its non-SSL version would fix the
problem, but would introduce mixed active content [1] which was removed
in this commit [2].

Note that while investigating the problem, I updated the FeedBurner feed
source from http://gofreerange.com/feed.xml ->
https://gofreerange.com/feed.xml, so now the sequence is:

https://feeds.gofreerange.com/GoFreeRangeBlog ->
https://1mcuj68.feedproxy.ghs.google.com/GoFreeRangeBlog ->
https://feeds.feedburner.com/GoFreeRangeBlog ->
https://gofreerange.com/feed.xml

Given that I don't think we ever really look at the analytics in
FeedBurner, nor do we use any of its other services (e.g. RSS to email),
and it looks as if Google is slowly letting it die, I think the best
option for now is to ditch FeedBurner and serve the feed directly from
our site. However, we should continue to support the original URL for
existing subscribers, so I think we want to end up with something like
this:

https://feeds.gofreerange.com/GoFreeRangeBlog ->
https://gofreerange.com/feed.xml

I think we'll need to do the following:

1. Update Rails app to serve
https://feeds.gofreerange.com/GoFreeRangeBlog from PagesController#feed

2. Add feeds sub-domain to Heroku app

3. Generate SSL certificate for feeds sub-domain (I think step 2 might
do this automatically)

4. Update DNS CNAME for feeds sub-domain to point to Heroku app instead
of FeedBurner

[1]: https://developer.mozilla.org/en-US/docs/Web/Security/Mixed_content#Mixed_active_content
[2]: https://github.com/freerange/site/commit/74ba8979b79b424b6a1233791d6b62d2900b7efd