Add wiki page with form to request invitation to S&T Slack
I plan to link to this from the show-and-tell-events snip page instead
of using the Slackin JavaScript. This should mean we can get rid of our
Slackin installation on Heroku.

I plan to use our existing Slack API token from our Slackin installation
on Heroku. However, it's worth noting this is probably a "legacy" token
and if we ever need to re-generate it, we'll probably have to jump
through some OAuth hoops.