Replace Slackin JS with link to new Slack invitation form
I haven't bothered re-implementing the counts of active/total users, but
this should be possible if we really care about it.

I also haven't bothered re-implementing the little Slack icon and
associated markup - I don't think it's worth the trouble.

I've removed the Slackin widget from the Show & Tell layout sidebar,
because it didn't look very good as just a link and in any case, I'm not
sure it's terribly useful there.