Use jquery-rails Gem
Instead of vendorising jquery source.

This upgrades jquery from 1.4.3 to 1.12.4.

This upgrade broke our footNoteLink plugin. The plugin scrolls back to
the link you clicked to get to the footnote. You can see it in action by
click on the return arrow next to one of the footlinks at the bottom of
https://gofreerange.com/show-and-tell-16.

The first error with footNoteLink was:

    Error: Syntax error, unrecognized expression: sup[id=fnref:3]

This was fixed by quoting the "fnref:3" string in the selector.

The next error was:

    TypeError: d.browser is undefined

This was fixed by upgrading jquery.scrollto to version 2.1.2. I
downloaded it from the releases page[1] and copied the minified version
to our vendor directory.

[1]: https://github.com/flesler/jquery.scrollTo/releases