Set Cache-Control header for asset pipeline assets
The RAILS_SERVE_STATIC_FILES env var is set in our Heroku environment so
that the Rails app serves asset pipeline assets which are compiled to
the public/assets directory.

This configuration change means that Rails adds a Cache-Control response
header for these assets with a 1 year max-age as recommended by Google's
Lighthouse auditing tool [1].

I _think_ there might be an issue with non-asset-pipeline assets living
under the public directory also having this response header added, but
I'm not sure. If it does turn out to be the case, we'll need to make
sure we rename any of these files if they ever change their content.

[1]: https://developers.google.com/web/tools/lighthouse/audits/cache-policy