Make nav bar position relative vs fixed
I'm not convinced that having a fixed nav bar was that useful and it has
always messed with the scrolling to anchors within the page.

The scrollTo jQuery plugin was used to fix this in the case of markdown
footnote back references, but not in other scenarios. And in any case, I
don't think using non-browser scrolling is very sensible.

This change allows me to remove the scrollTo jQuery plugin which is a
first step on being able to remove jQuery altogether.