Replace chronoClass jQuery plugin with custom server-side Ruby
This plugin was just being used to add a CSS class to the body element
on each page indicating the day of the week so that we can display a
different background on Thursdays (for some inexplicable reason!).

This is another step towards being able to get rid of jQuery altogether.