Add a "we are a coop" section to the Get in Touch section
I wanted a little bit of whitespace between the second h3 in this
sidebar (the new "we are a coop" header), so I've added some padding
to anything that's not the first h3 in the div.

I've used a unicode non-breaking hyphen `&#x2011;` entity to prevent
the word `co-operative` being broken over two lines.