Ignore all exceptions handled by Rails ActionDispatch::ShowExceptions
I can't see why we'd want to report any of the exceptions which
ActionDispatch::ShowExceptions rescues and converts into suitable HTTP
response codes using ActionDispatch::ExceptionWrapper [1]. This approach
is probably somewhat draconian and we might miss some important
exceptions. However, given the website has been around a long time and
we're pretty confident in its operation, I don't think this is a big
risk and the benefit is to improve the signal-to-noise ratio in our
exception monitoring.

This commit adds a before_process block to the Rollbar configuration so
that we only report exceptions which will not be handled by the Rails
middleware.

[1]: https://github.com/rails/rails/blob/v5.2.4.2/actionpack/lib/action_dispatch/middleware/exception_wrapper.rb#L9-L21