Remove Slack invitation form
We're no longer hosting Show & Tell events and the Slack API token is
about to be revoked due to inactivity [1]. Also this type of API token
is about to be phased out and the newer type would involve jumping
through a bunch of OAuth hoops. So overall it doesn't seem worth
maintaining this mechanism allowing new people to invite themselves to
the S&T Slack channel.

[1]: https://groups.google.com/a/gofreerange.com/d/msg/members/GQeSJd3StOw/qv1QTVgUAAAJ