Specify action_dispatch.cookies_same_site_protection = :lax
Previously the SameSite attribute was not set on our cookie. In most
recent browsers not having it set is equivalent to setting it to "lax",
so this shouldn't make any noticeably difference to users.

I suspect we could & should set the attribute to "strict", but that can
wait for now.