Don't bother to set action_controller.urlsafe_csrf_tokens
The last call to ActionView::Helpers::CsrfHelper::csrf_meta_tags was
removed in this commit [1] and we still don't have *any* forms, so this
isn't relevant.

We don't need to worry about forgetting to set it if/when we add a form,
because I'm planning to switch the Rails defaults to v6.1 in the near
future and that should enable this setting.

[1]: a5d9c6a34eb556568273c6a22d9b5edc4510c3a5