Fix non-breaking hyphens in Chrome
These HTML entities were not displaying correctly in Chrome.

To fix this I've replaced the non-breaking hyphens with normal hyphens
and wrapped the relevant words in span elements with the "white-space:
nowrap" CSS style applied. I'm not sure this is the best or most elegant
solution but it seems to work in Chrome, Firefox & Safari.