Use Rails 7 apply_stylesheet_media_default
This no longer adds `media="screen"` attribute to stylesheet link tags.
When there's no media attribute set the browser treats it the same as
being set to `all`. I suspect it doesn't matter whether it's set to
`screen` or `all` in our case so I'm happy to go with the new default.

There's some info about this change in
https://github.com/rails/rails/issues/41213.

Before

<link rel="stylesheet" href="/assets/application-ca84842787ea2b5de1cada11fbde0e5ae74f2220f20b2900d1c838c7028c821e.css" media="screen" />
<link rel="stylesheet" href="https://use.typekit.net/rny4upv.css" media="screen" />

After

<link rel="stylesheet" href="/assets/application-ca84842787ea2b5de1cada11fbde0e5ae74f2220f20b2900d1c838c7028c821e.css" />
<link rel="stylesheet" href="https://use.typekit.net/rny4upv.css" />