Remove unhelpful comments abour AWS SES in zone file
I've captured more detailed/useful info in comments directly on the
relevant DNS records in DNSimple.

In summary, we've used AWS SES for sending emails from
video@gofreerange.com (which is currently still setup in AWS, but not
being used) and (almost certainly) from mission-patch@gofreerange.com
(which is no longer being used since we switch to using Postmark).

This comment was annoying me when backing up our DNS records, because it
was the only thing preventing us from pasting in the whole exported zone
file.