Remove DNS records for AWS SES for mission-patch.com
I just removed the mission-patch.com AWS SES identity from the
`eu-west-1` region. These DNS records relate to that identity. This
identities was probably originally used to send emails from the Mission
Patch app. However, some time ago we switched to using Postmark to send
emails, so it is no longer needed and we can remove the associated DNS
records.

Note that I have left the `_dmarc.mission-patch.com` TXT record in
place, because (a) it doesn't seem specific to AWS SES (i.e. it _might_
relate to other email sending from this domain); and (b) the value is
different to that in the AWS console: "v=DMARC1; p=none;".