Run rails app:update & accept relevant changes
I've accepted the vast majority of the changes.

The only significant change that I did not accept was the removal of the
following line from `config/application.rb`:

    require "sprockets/railtie"

This line was added in this commit [1], because we're still indirectly
relying on sprockets to compile our stylesheets due to the dependency on
`sass-rails`.

[1]: https://github.com/freerange/site/commit/987f0657c1fd8a825664eb9c82831f073693bdf5