Update spec re requests for page in unknown format
This spec and the corresponding code in `PagesController#show` were
added in this commit [1].

The behaviour when a page is requested with an unknown format seems to
have changed slightly at least within the request spec in that an
`UnknownFormat` exception is no longer raised, but instead the response
has the 406 Not Acceptable status.

Previously, although the `UnknownFormat` exception was being raised in
the spec, it was being converted to a 406 Not Acceptable response in
middleware in production code. And so the outward facing behaviour in
production has not changed.

Thus it makes sense to just update the spec to assert for the new
behaviour.

[1]: https://github.com/freerange/site/commit/9abe7ef719e375bb9cabf96f29f861b7b6bf0b4a